Update README.md
I attached all of the graphs when I did my final commit on Nov 13. I spent over an hour making all of the graphs with Athena in the computer science center. I’m not sure what happened, there must have been a glitch with the README file (this has happened to me before in 112).